Fix project explorer select node resetting current page; #194 (#201)
This commit resolves an issue where selecting a node in the project
explorer would reset the currently selected page, which is
annoying when handling schemas with many nodes.

It accomplishes this by keeping a map of schema UUIDs to their
current pages in the explorer controller and the explorer
contents list.

In the contents list, this is necessary so that the
pagination directive can permanently store the page number in the
parent directive. Just storing it in the scope (the default
behaviour) was not sufficient - it seems that the page number
in the scope gets reset when overwriting the contents array entirely
in populateChildNodes(). Otherwise, the correct page gets retrieved,
but the pagination still shows the wrong page number.

In the explorer, this is necessary so that the correct page is
also requested from the API.